Move ORCA-debugging UDFs to a separate contrib module.
Some of the functions in funcs.cpp were only used by debugging/testing
code. They were included in the libgpoptudfs.so library, but it was built
in a strange way, from gpAux/Makefile. Move it to a contrib module, along
with the regression tests using it.

This commit leaves only the functions actually used by backend code in
funcs.cpp, and moves the debugging functions only used by the optional
UDFs into the contrib module.

NOTE: The regression tests don't actually work at the moment. But that's not
new, they didn't work before either. I don't have the patience to start
fixing them, so I'm just moving them to the new location in their current
form.

NOTE2: The new module is called orca_debug, but I think some test code
in GPORCA repository expects to find this in libgoptudfs.so. Will need
to fix the scripts there.
