+++ contrib/orca_debug/udf_input/exec02_const_text.mdp
