+++ contrib/orca_debug/udf_input/s02_const_string.xml
